Function that executes a function given as a parameter on each element of an array

+++ function_pointers/1-array_iterator.c
@@ -14,9 +14,10 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	size_t i;
 
 	i = 0;
+
 	if (action != NULL)
 	{
 		while (i < size)
